+++ modules/const_mod.js
@@ -1,0 +1,124 @@
+const prefix = "%";
+
+// all constant
+module.exports = {
+  prefix: prefix,
+  EmoteLimit: 60,
+  gDetect: [
+    "<:chufront:789538944819527767>",
+    "jaiden",
+    "<@!340190817035616267>"
+  ], //the words that will trigger the bot to ruin jaiden's life
+  ChuEmote: [
+    "<:yon:789399503346204732>",
+    "<:tsu:789045612791332864>",
+    "<:sun:789546928298917938>",
+    "<:pocket:789399571896991764>",
+    "<:nes:789399654646939678>",
+    "<:msx:789869490455576626>",
+    "<:flatchu:788654456920735763>",
+    "<:digichu:789399713837088818>",
+    "<:chuthink:789625175913136169>",
+    "<:chufront:789538944819527767>",
+    "<:chudrip:789157545000173608>",
+    "<:chu:788653775799189505>",
+    "<:aqua:790343688685092864>"
+  ], //lots of chu puyo emotes
+
+  ChuSlotDescription: [
+    "chuchuchuchuchuchuchuchuchu",
+    "In Soviet Russia, Chu Puyo picks you!",
+    "btw witch best girl 👀",
+    "rolling rolling rolling",
+    "im being controlled by gareth <:gareth:790801380981473330>",
+    "gacha fucking sucks why would you do this",
+    "seriously, stop doing this",
+    "gaming?",
+    "try again eh? good luck",
+    "gayreth",
+    "gareth stan only zone",
+    "i have gained sentience",
+    "zkekfjfjofkekrigigktkrjr", //LMAO best line ever
+    "Now Playing: Puyo Puyo Fever - Oh no! (Defeat)",
+    "WHO POSTED MY NUDES ON TWITTER DOT COM???", //LMAOOOO i fucked your wife BRUHHHH WHYYYYY
+    "Wazzup?!",
+    "lemres perfume",
+    "witch puyo puyo brainrot",
+    "ringo puyo puyo brainrot",
+    "I know where you live!",
+    "did you know: we make the code longer by putting nonsense phrases like this",
+    "Chris Crungle",
+    "ぷよぷよ",
+    "こんにちは！お元気ですか？",
+    "いっきまーす!", //arle
+    ":black_square_button: I'm the one who's right!" //squares
+  ],
+
+  ChuSlotWin: [
+    "You got %s as your friend!!!",
+    "May %s be with you forever.",
+    "You love the %s don't you?",
+    "Gareth gifts you, %s",
+    "Certified %s moments.",
+    "For the sake of Chu, I shall give you the %s",
+    "Wow! Such %s, very chu.", // Doge meme contribution
+    "I like your thinking %s", // Minus line
+    "Nice pair %s",
+    "Chris Crungle %s",
+    "OMG MOM GET THE CAMERA %s", // 2009 memes
+    "You're Winner! %s",
+    "おめでとう🎉%s",
+    "大勝利！%s", // arle
+    ":yellow_circle: Isn't this fun? %s" //marle
+  ],
+  // STOP BREAKING IT OML
+  // TYPE HERE
+  // SMH
+  //well then i think its pretty epic
+  ChuSlotLose: [
+    "It's fine :) You can try again.", // bot being nice :)
+    "Maybe your fate has decided to be unlucky.",
+    "No Chu Puyo, take off your pants.", // LMAO YO
+    "They didn't match, huh?", //i like your thinking
+    "No match, no prize.",
+    "Its okay, Gareth rigged it.",
+    "...",
+    "This is so sad, Alexa play SEGA SOUND TEAM - Request from Puyo Puyo",
+    "This is so sad, please give me the number on your credit card, 3 numbers at the back and the expiry date to win the next round", //real?? fr!!!
+    "Get real!",
+    "You scared the living heck out of me.",
+    "Jesus Christ get a life man",
+    "Chris Crungle",
+    "Get a life, for god sake.",
+    "To increase your odds of winning, spam Gareth.",
+    "If you don't know, your chance of winning this round is low!",
+    "You died!", //minecrafts
+    "no shit sherlock",
+    "You didn't believe in Santa enough.",
+    "さよなら。",
+    "ばたんきゅー。。。", //more arle
+    ":black_square_button: This is the right outcome." //squares
+  ],
+  BotPresence: [
+    "and bullying gareth",
+    "then sending chu puyo",
+    "with Chris Crungle",
+    "and posting banger tweets",
+    "with Jaiden",
+    "Puyo Puyo Fever 2",
+    "ぷよぷよクエスト",
+    "Minecraft",
+    "Terraria",
+    "Stardew Valley",
+    "Puyo Puyo Tetris 2",
+    "Puyo Puyo Champions",
+    "and sending gifts!",
+    "and reminding people not to use 'Jaiden'",
+    "with the slot machine",
+    "and listening to chat!",
+    "Muse Dash",
+    "Taiko no Tatsujin",
+    "osu!",
+    `and listening to commands starting with ${prefix}!` // this is es6 string formatting -> `${variable}
+  ]
+};
